Only canonicalize name if 8 or 16 characters was used
The apt_key2 canonicalize method was changing any name not 40 characters in length to the 40 character fingerprint.
This is helpful, but is not consistent with the old provider code that fails validation if the name is not 8, 16 or 40 characters long.
This commit makes apt_key2 consistent with that old behavior and leaves a TODO: item to restore canonicalization of any name length once we have finished demonstrating apt_key/apt_key2 equivalency.